Fix warnings EXIV2_TEAM_EXTRA_WARNINGS
Most of the warnings fixed were:
- Hiding of local variables
- Hiding of globals
- Useless casts
- Conversions between float/double

+++ src/jp2image.cpp
@@ -501,7 +501,7 @@ static void boxes_check(size_t b,size_t m)
                 enforce(box.length <= io_->size()-io_->tell() , Exiv2::kerCorruptedMetadata);
 
                 if (bPrint) {
-                    out << Internal::stringFormat("%8ld | %8ld | ", (size_t)(position - sizeof(box)),
+                    out << Internal::stringFormat("%8ld | %8ld | ", position - sizeof(box),
                                                   (size_t)box.length)
                         << toAscii(box.type) << "      | ";
                     bLF = true;
